+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - `--debug` didn't work because of a tiny error with `modules/log.py` (1.0.0 -> 1.0.1)
+- `config.py 1.0.1`: Various cleanup and fixes
  
 ## [2.0.0] - 2025-07-16